refactor: 시나리오 설정 유틸과 상수 분리
